pyproject.toml
[tool.poetry]
name = "drf-app"
version = "0.1.0"
description = ""
authors = ["Hotline Emu <11667661+hotline-emu@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.2"
pylint-django = "^2.6.1"
mypy = "^1.13.0"
django-stubs = "^5.1.1"
mypy-extensions = "^1.0.0"
djangorestframework-stubs = "^3.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests/test_*.py"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
strict = true
plugins = ["mypy_django_plugin.main"]
disable_error_code = ["attr-defined"]

[tool.django-stubs]
django_settings_module = "config.settings"
